+++ front/components/base/icomoon.component.tsx
@@ -58,6 +58,7 @@ export enum IcomoonIcons {
   mentionsLegales = "mentions-legales",
   politiquesConfidentialite = "politiques-confidentialite",
   glossaire = "glossaire",
+  stepParentheque = "step-parentheque",
   stepProjetParent = "step-projet-parent",
   stepConception = "step-conception",
   stepDebutDeGrossesse = "step-debut-de-grossesse",